Preset Manager fixes
- Enable buttons in right-click view
- Fix double value rounding in parameter/preset panels
- Fix issue with buttons getting disabled in Configuration Manager
- Fix issues with invalid data copied/pasted between parameters and
presets
- Fix duplicate save warnings on close of unsaved preset file

+++ Juce/ScopeSyncShared/Utils/BCMMath.h
@@ -72,7 +72,7 @@ namespace BCMMath
             return 0;
     }
 
-	inline double roundDouble(float num)
+	inline double roundDouble(double num)
     {
 		return (num > 0.0) ? floor(num + 0.5) : ceil(num - 0.5);
 	}
